Further tuning of color generation.  At present, CMY controls can be used to balance grayscale reasonably effectively.  The grayscale for the Stylus Photo 870/1270 was tuned with this.
Change use of floats to doubles for better accuracy (and because
floats are always passed as doubles in C, anyway).

Change GhostScript -sMediaSource to -sInputSlot to conform to PPD
standard.

Nominally support Stylus Color 880 and 980.

+++ escp2-unprint.c
@@ -166,11 +166,11 @@ void mix_ink(ppmpixel p, int c, unsigned int a) {
 
   /* this is pretty crude */
 
-  float ink[3];
-  float size;
+  double ink[3];
+  double size;
 
   if (pstate.dotsize&0x10) {
-    size=(float)a/3.0;
+    size=(double)a/3.0;
   } else {
     size=1.0;
   }
