+++ print-canon.c
@@ -19,18 +19,6 @@
  *   You should have received a copy of the GNU General Public License
  *   along with this program; if not, write to the Free Software
  *   Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
- *
- * Contents:
- *
- *   canon_parameters()     - Return the parameter values for the given
- *                            parameter.
- *   canon_imageable_area() - Return the imageable area of the page.
- *   canon_print()          - Print an image to a CANON printer.
- *   canon_write()          - Send 6-color graphics using compression.
- *
- * Revision History:
- *
- *   See ChangeLog
  */
 
 
@@ -805,7 +793,7 @@ canon_print(const printer_t *printer,		/* I - Model */
   int 		output_type = v->output_type;
   int		orientation = v->orientation;
   char          *ink_type = v->ink_type;
-  float 	scaling = v->scaling;
+  double 	scaling = v->scaling;
   int		top = v->top;
   int		left = v->left;
   int		y;		/* Looping vars */
